requirements.txt
Flask==0.12.1
google-cloud-storage==1.6.0
